+++ Sources/Parameters/Parameters.hpp
@@ -25,6 +25,7 @@
 #pragma once
 #define STONEYDSP_BIQUADS_PARAMETERS_HPP_INCLUDED
 
+#ifndef STONEYDSP_BIQUADS_HPP_INCLUDED
 #include <juce_audio_basics/juce_audio_basics.h>                                // depends: juce_core
 #include <juce_audio_processors/juce_audio_processors.h>                        // depends: juce_gui_extra, juce_audio_basics
 #include <juce_core/juce_core.h>                                                //
@@ -33,8 +34,9 @@
 #include <juce_graphics/juce_graphics.h>                                        // depends: juce_events
 #include <juce_gui_basics/juce_gui_basics.h>                                    // depends: juce_graphics, juce_data_structures
 #include <juce_gui_extra/juce_gui_extra.h>                                      // depends: juce_gui_basics
+#endif // STONEYDSP_BIQUADS_HPP_INCLUDED
 
-// #include "Biquads.hpp"
+// #include <Biquads.hpp>
 
 namespace StoneyDSP {
 /** @addtogroup StoneyDSP @{ */
