+++ Sources/Processor/Processor.hpp
@@ -25,6 +25,7 @@
 #pragma once
 #define STONEYDSP_BIQUADS_PROCESSOR_HPP_INCLUDED
 
+#ifndef STONEYDSP_BIQUADS_HPP_INCLUDED
 #include <juce_audio_basics/juce_audio_basics.h>                                // depends: juce_core
 #include <juce_audio_formats/juce_audio_formats.h>                              // depends: juce_audio_basics
 #include <juce_audio_processors/juce_audio_processors.h>                        // depends: juce_gui_extra, juce_audio_basics
@@ -35,6 +36,14 @@
 #include <juce_graphics/juce_graphics.h>                                        // depends: juce_events
 #include <juce_gui_basics/juce_gui_basics.h>                                    // depends: juce_graphics, juce_data_structures
 #include <juce_gui_extra/juce_gui_extra.h>                                      // depends: juce_gui_basics
+
+#include <stoneydsp_audio/stoneydsp_audio.h>
+#include <stoneydsp_core/stoneydsp_core.h>
+#endif // STONEYDSP_BIQUADS_HPP_INCLUDED
+
+// #include <Biquads.hpp>
+// #include <Parameters/Parameters.hpp>
+// #include <Wrapper/Wrapper.hpp>
 
 // #include "Biquads.hpp"
 // #include "Parameters.hpp"
